requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.35.10
botocore==1.35.10
cffi==1.17.0
cryptography==43.0.0
Django==4.2.15
django-cors-headers==4.4.0
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
ecdsa==0.19.0
jmespath==1.0.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
